requirements.txt
streamlit
PyPDF2
langchain
langchain-community
google-generativeai
langchain_google_genai
python-dotenv
faiss-cpu
